d_lambda change type from int to float

+++ main.py
@@ -15,7 +15,7 @@ eps = np.finfo(float).eps # epsilon for avoiding zero division
 parser = argparse.ArgumentParser(description='Normal Integration by solving Poisson Equation')
 parser.add_argument('normal', help='the path of normal map')
 parser.add_argument('-d', '--depth', default=None, help='the path of depth prior')
-parser.add_argument('--d_lambda', type=int, default=100, help='how much will the depth prior influence the result')
+parser.add_argument('--d_lambda', type=float, default=100, help='how much will the depth prior influence the result')
 parser.add_argument('-o', '--output', default='output', help='name of the output object and depth map')
 parser.add_argument('--obj', dest='write_obj', action='store_const',
                     const=True, default=False, help='write wavefront obj file, by default False')
